Add files via upload
This study demonstrates the application of genetic algorithms to portfolio optimization. The ultimate goal of portfolio optimization is to have the best return/risk ratio. This can be done by changing the weights assigned to each asset in the portfolio. For a given level of risk, the optimum portfolio will either have the lowest risk or the highest expected return.